+++ src/core/layout/gel-button.js
@@ -6,6 +6,30 @@
 import fp from "../../../lib/lodash/fp/fp.js";
 import * as signal from "../signal-bus.js";
 import * as GameSound from "../game-sound.js";
+import { gmi } from "../gmi/gmi.js";
+
+class Indicator extends Phaser.Sprite {
+    constructor(parent) {
+        super(parent.game, 0, 0, assetPath({ key: "notification", isMobile: parent._isMobile }));
+        this.parent = parent;
+        this.scale = { x: 0, y: 0 };
+        this.anchor.set(0.5, 0.5);
+        parent.game.add.tween(this.scale).to({ x: 1, y: 1 }, 500, Phaser.Easing.Bounce.Out, true, 1000);
+        parent.addChild(this);
+        this.resize();
+    }
+
+    resize() {
+        this.position.x = this.parent.width / 2;
+        this.position.y = this.parent.height / -2;
+        this.animations.sprite.loadTexture(assetPath({ key: "notification", isMobile: this.parent._isMobile }));
+    }
+}
+
+const noIndicator = {
+    resize: () => {},
+    destroy: () => {},
+};
 
 export class GelButton extends Phaser.Button {
     constructor(game, x, y, metrics, config) {
@@ -24,6 +48,8 @@ export class GelButton extends Phaser.Button {
         this.positionOverride = config.positionOverride;
         this.animations.sprite.anchor.setTo(0.5, 0.5);
         this.setHitArea(metrics);
+        this.indicator = noIndicator;
+        this.setIndicator();
     }
 
     setHitArea(metrics) {
@@ -43,6 +69,14 @@ export class GelButton extends Phaser.Button {
         this._isMobile = metrics.isMobile;
         this.animations.sprite.loadTexture(assetPath({ key: this._id, isMobile: metrics.isMobile }));
         this.setHitArea(metrics);
+
+        this.indicator.resize();
+    }
+
+    setIndicator() {
+        this.indicator.destroy();
+        const show = this._id === "achievements" && gmi.achievements.unseen;
+        this.indicator = show ? new Indicator(this) : noIndicator;
     }
 }