+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 | Version | Description |
 |---------|-------------|
+| 1.0.15 | |
 | | Add achievement button config for results screen. . |
 | | Add per screen button overrides. |
 | | Update webpack config so bowser package is run through babel. |